Cleaned up namespaces, renamed database access to UserRepo, fixed small mistakes
