--- Utilities/DatabaseAccess.cs
@@ -1,418 +1,0 @@
-﻿using System.Data;
-using System.Linq.Expressions;
-using System.Reflection;
-using MySql.Data.MySqlClient;
-using Serilog;
-using neismesk.Models;
-using static System.Net.Mime.MediaTypeNames;
-using neismesk.ViewModels.Item;
-using System.Data.Common;
-
-namespace neismesk.Utilities
-{
-    /// <summary>
-    /// This class provides access to the MySQL database.
-    /// </summary>
-    public class DatabaseAccess
-    {
-        private Serilog.ILogger _logger;
-        private readonly string _connectionString;
-
-        public DatabaseAccess()
-        {
-            CreateLogger();
-            
-            _connectionString = Environment.GetEnvironmentVariable("DATABASE_CONN_STRING");
-        }
-
-        public DatabaseAccess(string connectionString)
-        {
-            CreateLogger();
-            
-            _connectionString = connectionString;
-        }
-
-        private void CreateLogger()
-        {
-            _logger = new LoggerConfiguration()
-                .WriteTo.Console()
-                .CreateLogger();
-        }
-
-        private MySqlConnection GetConnection()
-        {
-            return new MySqlConnection(_connectionString);
-        }
-
-        /// <summary>
-        /// Returns data as a List of objects.
-        /// </summary>
-        /// <typeparam name="T">The type of the objects to return</typeparam>
-        /// <typeparam name="U">The type of the parameters used in the SQL query</typeparam>
-        /// <param name="sql">The SQL query to be executed</param>
-        /// <param name="parameters">The parameters used in the SQL query</param>
-        /// <returns>List of objects of type T</returns>
-        public async Task<List<T>> LoadData<T, U>(string sql, U parameters)
-        {
-            using MySqlConnection connection = GetConnection();
-            using MySqlCommand command = new MySqlCommand(sql, connection);
-            AddParameters(command, parameters);
-            
-            await connection.OpenAsync();
-            try
-            {
-                using MySqlDataReader reader = (MySqlDataReader)await command.ExecuteReaderAsync();
-
-                List<T> result = new List<T>();
-
-                while (await reader.ReadAsync())
-                {
-                    T row = Activator.CreateInstance<T>();
-                    for (int i = 0; i < reader.FieldCount; i++)
-                    {
-                        PropertyInfo property = row.GetType().GetProperty(reader.GetName(i));
-                        if (property != null && !reader.IsDBNull(i))
-                        {
-                            property.SetValue(row, reader.GetValue(i), null);
-                        }
-                    }
-                    result.Add(row);
-                }
-                return result;
-            }
-            catch (Exception ex)
-            {
-                _logger.Error(ex, "Error loading data from database!");
-                return null;
-            }
-        }
-
-        /// <summary>
-        /// Returns query results as a DataTable.
-        /// </summary>
-        /// <param name="sql">The SQL query to be executed</param>
-        /// <param name="parameters">List of parameters to append to the query</param>
-        /// <returns>DataTable containing the query results</returns>
-        public async Task<DataTable> LoadData<U>(string sql, U parameters)
-        {
-            try
-            {
-                using (var connection = new MySqlConnection(_connectionString))
-                {
-                    using (var command = new MySqlCommand(sql, connection))
-                    {
-                        AddParameters(command, parameters);
-
-                        await connection.OpenAsync();
-                        var dataTable = new DataTable();
-                        using (var dataAdapter = new MySqlDataAdapter(command))
-                        {
-                            dataAdapter.Fill(dataTable);
-                        }
-                        return dataTable;
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                _logger.Error(ex, "Error loading data from database!");
-                return null;
-            }
-        }
-
-        /// <summary>
-        /// Returns query results as a DataTable.
-        /// </summary>
-        /// <param name="sql">The SQL query to be executed</param>
-        /// <returns>DataTable containing the query results in plain form</returns>
-        public async Task<DataTable> LoadData(string sql)
-        {
-            try
-            {
-                using (var connection = new MySqlConnection(_connectionString))
-                {
-                    using (var command = new MySqlCommand(sql, connection))
-                    {
-                        await connection.OpenAsync();
-                        var dataTable = new DataTable();
-                        using (var dataAdapter = new MySqlDataAdapter(command))
-                        {
-                            dataAdapter.Fill(dataTable);
-                        }
-                        return dataTable;
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                _logger.Error(ex, "Error loading data from database!");
-                return null;
-            }
-        }
-
-        /// <summary>
-        /// Executes a query that modifies data in the database.
-        /// </summary>
-        /// <typeparam name="U">Type of parameter object</typeparam>
-        /// <param name="sql">SQL statement to execute</param>
-        /// <param name="parameters">Object containing parameter values</param>
-        public async Task<bool> SaveData<U>(string sql, U parameters)
-        {
-            try
-            {
-                using MySqlConnection connection = GetConnection();
-                using MySqlCommand command = new MySqlCommand(sql, connection);
-                AddParameters(command, parameters);
-
-                await connection.OpenAsync();
-                await command.ExecuteNonQueryAsync();
-
-                return true;
-            }
-            catch (Exception ex)
-            {
-                _logger.Error(ex, "Error saving data to database!");
-                return false;
-            }
-        }
-
-		/// <summary>
-		/// Executes a query that modifies data in the database and return the id of the modified entry.
-		/// </summary>
-		/// <typeparam name="U">Type of parameter object</typeparam>
-		/// <param name="sqlForInsert">SQL statement to execute</param>
-		/// <param name="parameters">Object containing parameter values</param>
-		public async Task<int> SaveDataGetId<U>(string sqlForInsert, string sqlForId, U parameters)
-		{
-			try
-			{
-				using MySqlConnection connection = GetConnection();
-				using MySqlCommand command = new MySqlCommand(sqlForInsert, connection);
-				AddParameters(command, parameters);
-
-				await connection.OpenAsync();
-				await command.ExecuteNonQueryAsync();
-
-                command.CommandText = sqlForId;
-
-                int id = Convert.ToInt32(await command.ExecuteScalarAsync());
-
-				return id;
-			}
-			catch (Exception ex)
-			{
-				_logger.Error(ex, "Error saving data to database!");
-				return -1;
-			}
-		}
-
-		/// <summary>
-		/// Adds parameters to a database command object.
-		/// </summary>
-		/// <typeparam name="U">Type of parameter object</typeparam>
-		/// <param name="command">Command object to add parameters to</param>
-		/// <param name="parameters">Object containing parameter values</param>
-		private void AddParameters<U>(IDbCommand command, U parameters)
-        {
-            if (parameters != null)
-            {
-                foreach (PropertyInfo property in parameters.GetType().GetProperties())
-                {
-                    IDbDataParameter parameter = command.CreateParameter();
-                    parameter.ParameterName = property.Name;
-                    parameter.Value = property.GetValue(parameters) ?? DBNull.Value;
-                    command.Parameters.Add(parameter);
-                }
-            }
-        }
-
-        /// <summary>
-        /// Executes a query that modifies data in the database.
-        /// </summary>
-        /// <typeparam name="U">Type of parameter object</typeparam>
-        /// <param name="sql">SQL statement to execute</param>
-        /// <param name="parameters">Object containing parameter values</param>
-        public async Task<bool> InsertImages(ItemModel item)
-        {
-            try
-            {
-                using MySqlConnection connection = GetConnection();
-                await connection.OpenAsync();
-
-                foreach (IFormFile image in item.Images)
-                {
-                    byte[] data = await ImageUtilities.ResizeCompressImage(image, 640, 480);
-
-                    using MySqlCommand command = new MySqlCommand(
-                        "INSERT INTO images (img, fk_ad) VALUES (@image, @fk_ad)", connection);
-
-                    // Add parameters
-                    command.Parameters.AddWithValue("@image", data);
-                    command.Parameters.AddWithValue("@fk_ad", item.Id);
-
-                    await command.ExecuteNonQueryAsync();
-                }
-
-                return true;
-            }
-            catch (Exception ex)
-            {
-                _logger.Error(ex, "Error saving images to database!");
-                return false;
-            }
-        }
-
-        public async Task<List<ItemImageViewModel>> GetImage (int id)
-        {
-            List<ItemImageViewModel> images = new List<ItemImageViewModel>();
-            try
-            {
-                using MySqlConnection connection = GetConnection();
-                await connection.OpenAsync();
-
-                using MySqlCommand command = new MySqlCommand(
-                    "SELECT img_id, img FROM images WHERE fk_ad=@id", connection);
-                command.Parameters.AddWithValue("@id", id);
-
-                using DbDataReader reader = await command.ExecuteReaderAsync();
-
-                while (await reader.ReadAsync())
-                {
-                    int dataLength = (int)reader.GetBytes(1, 0, null, 0, int.MaxValue);
-                    byte[] imageData = new byte[dataLength];
-                    reader.GetBytes(1, 0, imageData, 0, dataLength);
-
-                    ItemImageViewModel image = new()
-                    {
-                        Id = reader.GetInt32("img_id"),
-                        Data = imageData,
-                    };
-                    images.Add(image);
-                }
-                return images;
-            }
-            catch (Exception ex)
-            {
-                _logger.Error(ex, "Error getting images from the database!");
-                return images;
-            }
-        }
-
-        /// <summary>
-        /// Executes a query that modifies data in the database.
-        /// </summary>
-        /// <typeparam name="U">Type of parameter object</typeparam>
-        /// <param name="sql">SQL statement to execute</param>
-        /// <param name="parameters">Object containing parameter values</param>
-        public async Task<bool> InsertQuestions(ItemModel item)
-        {
-            try
-            {
-                using MySqlConnection connection = GetConnection();
-                await connection.OpenAsync();
-
-                foreach (string question in item.Questions)
-                {
-                    using MySqlCommand command = new MySqlCommand(
-                        "INSERT INTO questions (question_text, fk_ad) VALUES (@question, @fk_ad)", connection);
-
-                    // Add parameters
-                    command.Parameters.AddWithValue("@question", question);
-                    command.Parameters.AddWithValue("@fk_ad", item.Id);
-
-                    await command.ExecuteNonQueryAsync();
-                }
-
-                return true;
-            }
-            catch (Exception ex)
-            {
-                _logger.Error(ex, "Error saving questions to database!");
-                return false;
-            }
-        }
-
-        public async Task<List<ItemQuestionViewModel>> GetQuestions(int itemId)
-        {
-            List<ItemQuestionViewModel> questions = new List<ItemQuestionViewModel>();
-            try
-            {
-                using MySqlConnection connection = GetConnection();
-                await connection.OpenAsync();
-
-                using MySqlCommand command = new MySqlCommand(
-                    "SELECT question_text, id FROM questions where fk_ad=@itemId", connection);
-                command.Parameters.AddWithValue("@itemId", itemId);
-
-                using DbDataReader reader = await command.ExecuteReaderAsync();
-                while (await reader.ReadAsync())
-                {
-                    int id = reader.GetInt32("id");
-                    string text = reader.GetString("question_text");
-                    ItemQuestionViewModel question = new ItemQuestionViewModel { Id = id, Question = text };
-                    questions.Add(question);
-                }
-
-                return questions;
-            }
-            catch (Exception ex)
-            {
-                _logger.Error(ex, "Error getting questions from database!");
-                return questions;
-            }
-        }
-
-        public async Task<List<ItemViewModel>> Search(string searchWord)
-        {
-            List<ItemViewModel> foundItems = new List<ItemViewModel>();
-            try
-            {
-                using MySqlConnection connection = GetConnection();
-                await connection.OpenAsync();
-
-                using MySqlCommand command = new MySqlCommand(
-                    "SELECT id, name, description, fk_user FROM ads " +
-                    "WHERE name LIKE CONCAT('%', @searchWord, '%') " +
-                    "OR description LIKE CONCAT('%', @searchWord, '%')", connection);
-                command.Parameters.AddWithValue("@searchWord", searchWord);
-
-                using DbDataReader reader = await command.ExecuteReaderAsync();
-                while (await reader.ReadAsync())
-                {
-                    ItemViewModel item = new()
-                    {
-                        Id = reader.GetInt32("id"),
-                        Name = reader.GetString("name"),
-                        Description = reader.GetString("description"),
-                        UserId = reader.GetInt32("fk_user"),
-                        Images = await GetImage(reader.GetInt32("id"))
-                    };
-                    foundItems.Add(item);
-                }
-
-                return foundItems;
-            }
-            catch (Exception ex)
-            {
-                _logger.Error(ex, "Error getting items by search word from database!");
-                return foundItems;
-            }
-        }
-
-        public bool TestConnection()
-        {
-            // Return true if the connection is successful
-            using MySqlConnection connection = GetConnection();
-            connection.Open();
-            // Check if the connection was  oppened successfully
-            if (connection.State == ConnectionState.Open)
-            {
-                return true;
-            }
-            else
-            {
-                return false;
-            }
-        }
-    }
-}
